REF: collect relevant comparisons is now make_run_page job.  RunCheck takes a list, while DataWidgets take a single dataclass
